// frontend/src/types/generated/ProviderRequestQuestion.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ProviderRequestOption } from "./ProviderRequestOption.js";

export type ProviderRequestQuestion = { id: string, header: string, question: string, options: Array<ProviderRequestOption>, multiple: boolean, is_other: boolean, is_secret: boolean, };